Added logic to automatically charge the inverter based on expected power consumption and generation forecasts. Imported necessary modules, adjusted data types, and utilized the `pause` library to control the timing of the charging process.

+++ source/inverter.py
@@ -44,7 +44,7 @@ class Inverter(LoggerMixin):
         self.log.info("Successfully set new operation mode")
 
     @staticmethod
-    def calculate_necessary_duration_to_charge(amount_of_energy: float) -> float:
+    def calculate_necessary_duration_to_charge(amount_of_energy: float) -> int:
         """
         Calculates how long the inverter needs to be fully charged.
